feat: convert to an npm module
BREAKING CHANGE: now an npm module and not Bower compatible.

- Switch to Angular commit style
- Setup semantic release
- Convert to ES2015 with babel
