+++ .eslintrc
@@ -1,0 +1,18 @@
+{
+  "parser": "babel-eslint",
+  "extends": [
+    "standard"
+  ],
+  "env": {
+    "jasmine": true
+  },
+  "rules": {
+    // overrides of the standard style
+    "curly": [2, "all"],
+    "indent": [2, 4],
+    "max-len": [2, 100, 4],
+    "semi": [2, "always"],
+    "space-before-function-paren": [2, {"anonymous": "always", "named": "never"}],
+    "wrap-iife": [2, "outside"]
+  }
+}
