--- .gitattributes
@@ -1,2 +1,0 @@
-# Automatically normalize line endings for all text-based files
-* text=auto
