+++ .gitignore
@@ -1,2 +1,2 @@
-bower_components
 node_modules
+dist/
