+++ .travis.yml
@@ -1,6 +1,20 @@
+sudo: false
 language: node_js
+cache:
+  directories:
+    - node_modules
+notifications:
+  email: false
 node_js:
-  - "0.10"
+  - '4'
+before_install:
+  - npm i -g npm@^2.0.0
 before_script:
+  - npm prune
   - export DISPLAY=:99.0
   - sh -e /etc/init.d/xvfb start
+after_success:
+  - npm run semantic-release
+branches:
+  except:
+    - "/^v\\d+\\.\\d+\\.\\d+$/"
