--- bower.json
@@ -1,21 +1,0 @@
-{
-  "name": "flight-with-child-components",
-  "version": "0.2.4",
-  "main": "lib/flight-with-child-components.js",
-  "dependencies": {
-    "flight": ">=1.1.4"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "jasmine-flight": "~3.0.0"
-  },
-  "ignore": [
-    ".*",
-    "CONTRIBUTING.md",
-    "CHANGELOG.md",
-    "test",
-    "package.json",
-    "karma.conf.js",
-    "Gruntfile.js"
-  ]
-}
