+++ config/webpack.config.test.js
@@ -1,0 +1,5 @@
+var baseConfig = require('./webpack.config');
+
+module.exports = Object.assign(baseConfig, {
+    devtool: 'inline-source-map'
+});
